+++ src/main/java/com/example/travelhana/Controller/UserController.java
@@ -109,5 +109,4 @@ public class UserController {
         return ResponseEntity.ok(tokens);
     }
 
-
 }
