+++ src/main/java/com/example/travelhana/Service/Implement/CustomUserDetailsImpl.java
@@ -1,5 +1,6 @@
-package com.example.travelhana.Service;
+package com.example.travelhana.Service.Implement;
 
+import com.example.travelhana.Service.CustomUserDetails;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
