+++ src/main/java/com/example/travelhana/Service/UserService.java
@@ -19,5 +19,7 @@ public interface UserService {
     void updateRefreshToken(String username, String refreshToken);
 
     Map<String,String> refresh(String refreshToken);
-    User getUser(String header);
+
+    User getUserByAccessToken(String header);
+
 }
